refactor(InputView): 자동차 대수 입력 제거 후 이름을 입력 받도록 수정
