+++ src/main/java/racinggame/exception/InputValueException.java
@@ -1,0 +1,8 @@
+package racinggame.exception;
+
+public class InputValueException extends RuntimeException {
+
+    public InputValueException(String message) {
+        super(message);
+    }
+}
